New prefixes and better quoting
Changed the prefixes to /Library/Gnucash-2.2.9 and 2.3.x. This is both the mac standard place for this and is user-writable which gets rid of the need for Applescript to authenticate. 

Fixed up all of the unquoted shell variables (or added a comment where the variable should stay unquoted).

Betwixt the two changes, the problem of paths-with-spaces should be resolved.