Bug 799350 - Opening MySQL database fails in 5.7 but works in 5.6
Mariadb-connector-C has a new default parameter that requires the
server to have a valid TLS certificate. Turn that off.

Reviewed the other options in CMakeLists.txt and the cmake output and
added some other options to silence warnings, use the jhbuild-built zlib,
and get the install path right so that it doesn't have to be fixed after
building.